redesign get measurements
