+++ netatmo-get-stations-data.html
@@ -42,12 +42,12 @@
         <input type="text" id="node-input-creds" placeholder="Add netatmo creds">
     </div>
     <div class="form-row">
-		<label for="node-config-input-deviceId" title="id of device"><i class="fa fa-home"></i> Device ID</label>
+		<label for="node-input-deviceId" title="id of device"><i class="fa fa-home"></i> Device ID</label>
 		<div style="display: inline-block; position: relative; width: 70%; height: 20px;">
             <div style="position: absolute; left: 0px; right: 40px;">
-                <input type="text" id="node-config-input-deviceId" placeholder="70:ee:....." style="width: 100%"/>
+                <input type="text" id="node-input-deviceId" placeholder="70:ee:....." style="width: 100%"/>
             </div>
-            <a id="node-config-input-deviceId-scan" class="editor-button" style="position: absolute; right: 0px; top: 0px;">
+            <a id="node-input-deviceId-scan" class="editor-button" style="position: absolute; right: 0px; top: 0px;">
                 <i class="fa fa-search"></i>
             </a>
         </div>
